// src/app/modules/business-common/src/lib/components/journey-status-tag/enums/translation-keys.enum.ts
export enum TranslationKeys {
  Journey_Status = 'Journey.Status',
  Journey_Status_KeyNode = 'Journey.Status.',
  Journey_Status_Cancelled = 'Journey.Status.Cancelled',
  Journey_Status_Confirmed = 'Journey.Status.Confirmed',
  Journey_Status_Delayed = 'Journey.Status.Delayed',
  Journey_Status_Departed = 'Journey.Status.Departed',
  Journey_Status_Diverted = 'Journey.Status.Diverted',
  Journey_Status_Landed = 'Journey.Status.Landed',
  Journey_Status_OnRoute = 'Journey.Status.OnRoute',
  Journey_Status_OnTime = 'Journey.Status.OnTime',
  Journey_Status_Returned = 'Journey.Status.Returned',
}